+++ lib/Cronjob/LogDelete.php
@@ -1,12 +1,22 @@
 <?php
 
-class rex_cronjob_log_delete extends rex_cronjob
+namespace FriendsOfRedaxo\ConsentManager\Cronjob;
+
+use Exception;
+use rex;
+use rex_addon;
+use rex_cronjob;
+use rex_i18n;
+use rex_sql;
+
+class LogDelete extends rex_cronjob
 {
     public function execute()
     {
         if (rex_addon::get('consent_manager')->isAvailable()) {
 
             try {
+                /** TODO: Umstellen auf nur rex_sql-Funktionen */
                 $sql = rex_sql::factory()->setQuery('DELETE FROdM ' . rex::getTable('consent_manager_consent_log') . ' WHERE createdate < DATE_SUB(NOW(), INTERVAL ' . (int) trim($this->getParam('days')) .' DAY)');
                 $noDeleted = $sql->getRows();